proxy/requirements.txt
anyio==3.6.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.91.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
jsonify==0.5
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pydantic==1.10.4
PyJWT==2.6.0
pytest==7.2.1
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.24.0
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.20.0
